requirements.txt
torch
torchaudio
styletts2
faster-whisper
ffmpeg-python
pydub
fastapi
uvicorn
numpy
pyyaml
librosa
phonemizer
nltk
soundfile
transformers
